fix asserts; move asserts to helpers
