+++ tests/test_cli.py
@@ -6,7 +6,7 @@ from conda.cli.common import arg2spec, spec_from_line
 
 from conda.compat import text_type
 
-from .helpers import capture_json_with_argv
+from .helpers import capture_json_with_argv, assert_in
 
 
 class TestArg2Spec(unittest.TestCase):
@@ -142,7 +142,7 @@ class TestJson(unittest.TestCase):
                 'envs_dirs', 'is_foreign', 'pkgs_dirs', 'platform',
                 'python_version', 'rc_path', 'root_prefix', 'root_writable')
         for key in keys:
-            self.assertIn(key, res)
+            assert_in(key, res)
 
         res = capture_json_with_argv('conda', 'info', 'conda', '--json')
         self.assertIsInstance(res, dict)
